Add first version of a documentation
This has the same structure as the
trustlines-protocol/www.trustlines.foundation repository. There is just
a short introducing root README which links to the developer file.
The developer README has been created to have a valid link. This
repository hasn't any package.json with script targets to document. This
will be hopefully added soon. At least the release/review process will
be establishes similar to the former linked website.

Related to: trustlines-network/project#798